Introduce Pulse and PulseToggle prefabs
Pulse: get() is true on the rising edge of
the input.

PulseToggle: maintains a state that alternates
on the rising edge of the input. on() is
true on the rising edge when state is true.
off() is true on the rising edge when state
is false.

Toggle has also been reimplemented in terms
of Pulse.