Rename Controller -> Coordinator, make macro-ish
The run() method of a Coordinator now returns a boolean
indicating whether the Coordinator should keep running.

The start() method sets the Coordinator in the running state
and calls its begin() hook.

The execute() method calls the Coordinator's run() hook if
it's currently running. If the run() hook returns false,
the Coordinator exits the running state and its end() hook
is called.

The stop() method exits the running state and calls the
Coordinator's end() hook if it was previously in the
running state.