Add proper driving to teleop
This include Arcade, Teleop, and Dynamic Drive, choosable from the dashboard.

The Xbox is no longer public static. That was a workaround last year to implement
joystick inversion. With the new framework, the inversion can be done in TeleopMode,
obviating the need to access it from another class.

TODOs: Expose drive mode info and inversion info on Dashboard.
TODOs: Implement pneumatic shifting